BUG(438): Don't die when wakeup pipe couldn't be allocated. Fix potential deadlock.
Moves allocation of wakeup pipe out of client thread so it can be error-
handled. It also contains a fix for a potential deadlock; the pipe wasn't
set in non blocking mode, so if the client thread was very busy the
pipe could fill up and cause xmms2d to deadlock. Only write one byte to the
wake up pipe, there is absolutly on point in writing two bytes.