BUG(462): musepack ape-tag portability
The portability problem was casting chunks of data to structs
which would probably not work everywhere<tm>. With this change
a lot of stuff could be removed so this patch also cleans up
incompatible/unused stuff.

The musepack plugin now checks if the tag is valid before checking
tag length, something that was done inside the apev2 parser before.

There was also a use of malloc instead of g_new in the musepack.c.