BUG(1192): Include the tutorials.
This is done using git submodules. To make things work do

  git submodule init   # add the submodule url to .git/config
  git submodule update # clone the submodule repo and check it out

The main repo stores the commit id of the submodule that it references
and that will be checked out with git submodule update. If changes are
made to the submodule repository the main repo needs to be told to
reference the updated submodule. That's done by

  pushd doc/tutorial
  git checkout master
  git pull
  popd
  git add doc/tutorial
  git commit

Gotchas:

* It's necessary to publish the changes to the submodule before pushing
  the change to the superproject that references it. If you don't others
  won't be able to clone.

* git submodule update isn't safe if you've made changes within the
  submodule. They will be silently overwritten.