BUG(2020): Slimmed down xmms_object_t.
Removed the large arrays for the signals and commands and replaced
them by search trees. That way we'll only allocate memory for those
signal lists and commands if we actually need them.

In a quick test, this resulted in a ~162 KB memory win.

While implementing this, I noticed possible memory leaks in the
command handling code (see the FIXME).