BUG(2059): Free xmmsv_t instances after we sent them over the wire.
Before xmmsv we needed to make sure we don't destroy GTrees,
GLists etc, so we used to only call g_free on the xmmsv_cmd_value_t.
The xmmsv_t's that we created don't know about the GLib objects they
are populated from, so we can just unref the xmmsv_t unconditionally
now.