OTHER: Reworked IPC method IDs.
Before we used to just have one big enum of all IPC methods for all
objects. Now the methods are split up for each object -- the first
method for any object will have the same ID now.
This makes describing the objects and methods in XML (or whatever)
easier, hopefully.