FEATURE(2262): Working trie-based completion of commands and subcommands (not arguments, and only in readline mode).
Initially based on a commit by AnthonyG, although not sure
any code survived the porting to trie-based completion.

Completion notably works in auto-complete mode and with aliases.
All of this only works in readline/shell-mode, obviously.
There is no completion for arguments yet.

An optional completions argument has been added to the
command_trie_find function, so completions are only computed
on request.

Still needs work to support completion for various kind of
command-specific arguments, and to expose the completion through
the inline interface (so that e.g. bash-completion can use it).

Ah and I need to write automated tests for this stuff, seriously.