BUG(2360): Link tests against ncurses if installed
On some distributions cunit is compiled against ncurses. On these, programs
using cuint must also be linked against ncurses. If not linking them succeeds
but they cannot be executed.
So this patch always links the testprograms against cunit.
The alternative would be to add a runtime test in the configure step, which
will cause problems when cross-compiling.