BUG(2410): Graceful shutdown of the glib connector.
This patch increfs the xmmsc_connection_t when starting
the main loop to prevent watcher->conn from ever being
freed while the loop is running. The second part is that
source id returned from the watcher is kept so that it
can be used to remove the polling mechanism on shutdown.