OTHER: Use weakrefs on conn instance in results.
Incrementing the reference count on the connection in a result
instance creates a reference loop when the result is queued for
asynchronous handling in the ipc layer (e.g. broadcasts, which remain
indefinitely in the queue). The consequence is that when the mainloop
exits, the client doesn't known if the results queue is empty. It
means that once the connection is released, the connection keeps
indirect references to itself (through the queued results), leading to
a memory leak if the application doesn't use a workaround by tracking
itself all the results it creates and remove them explicitely from the
queue before releasing the connection.

This patch replaces the ref by a weak ref: results don't hold a
reference on the connection anymore, but the pointer to the connection
is set to NULL once it is destroyed. The change has no implications on
clients that don't try to use results after the connection has been
released (it should never happen anyway). Clients which try and do
that will get a warning and the action will be ignored.