BUG(2007): Prevent most jack output underruns and failures
This patch makes use of the xmms_output_bytes_available function to
check if enough bytes are in the output buffer to run the jack callback.
If not it will fill the buffer with silence. This fixes most but not all
cases of jack underuns and hangups.

Increase the buffersize to handle up to 4096 frame periods in one
read. Start the port numbers at one instead of zero as starting at
one is generally convention. When an underrun happens, count it and
print it.