OTHER: Let playlist handle adding of entries.
For some weird reason the medialib was responsible for adding
entries to the playlist after it had added them to the medialib.

This removes the last medialib<->playlist coupling and medialib
is now free to be constructed outside the playlist object.

A future improvement is to make the playlist add the collection
in one transaction, which could be reused by a lot of other
playlist functions.