OTHER: Fix constness warning for rl_completer_word_break_characters
This is a weird global string defined in readline. Unlike the
rest of the global strings it has its const modifier commented
out:

extern /*const*/ char *rl_completer_word_break_characters;

Reason for this is unknown, but readline itself does not
free nor manipulate the string, so it's safe to cast a const
string to it.

For more details, check readline source:

 complete.c: _rl_find_completion_word
 readline.c: readline_initialize_everything