OTHER: Fix pointer to released memory.
PyBytes_AsString() access to an internal buffer. In some cases,
code generated by cython would DECREF the object passed in before
the pointer is used. We now explicitely hold a ref to the object
and implicitely call PyBytes_AsString() via an explicit cast <char *>

TODO: Check if the implicit call works in cython 0.13 and 0.14