OTHER: Strip comments in generated cython files
Cython add code context in a comment for each line of the source code.
It leads to oversized generated files and each change in source code
modify at least 10 lines in generated files.

Since Cython doesn't provide an option to disable thoses, we remove
them when writting pre-generated files in the source tree.