BUG(2521): Limit non-compilations to user query.
The commit to #2519 created a complement of:

Match(userquery, compilation=1)

...which doesn't mean the part of the user query which is
not compilations, but instead the whole medialib minus the
compilation part of the user query, which is obviously not
correct.

This patch applies intersection between the user query and
the complement of Match(userquery, compilation=1), which
equals all albums in the user query which is not compilations.