OTHER: Use objects generators instead of cstlib
Symbol from staticlibs are not included if not used in the target.
This is not a problem when linked against from an application (xmms2d),
but it is on shared libraries (libxmmsclient) that are expected to
expose thoses symbols anyway.

waf provides a way to transparently replace thoses staticlibs by
groups of 'objects' which do not suffer from the lazy imports.