OTHER: Add a script that generates audio files of different formats.
This might be the start of some kind of integration test where we actually
verify that we decode stuff correctly, have gapless playback etc. For now
it's just a nice way of not having to figure out where to find files of
odd format X, Y, or Z. It's not full coverage of what we support, but it's
at least a start.