BUG(2576): Upgrade waf to 1.8.11.
Fixes Ruby 2.2 config check that Fedora and Debian has patches
for since a while back.

Due to changes in waf, 'uselib' dependencies no longer work as
they used to. The benefit of this is that the workaround, localdeps
tool, adds additional convenience so that we can properly pull in
'uselib' dependencies via 'use' parameters, and thus remove a lot
of 'uselib' parameters across the build scripts.