OTHER: Add a tool to verify build artifacts.
Intended use is to stash away two _build_ directories in, lets say,
"build1" and "build2". These two directories are passed to the script:

utils/compare-builds.py build1 build2

...which will output what files differ in checksum, and how many bytes
off the pairs are if there is a difference, also lists missing files
in each of the directories.

In the case of gzip files (man-pages) the decompressed content is
checksummed instead as the gzip header contains timestamp details.