OTHER: Add default-hint tag for methods arguments
It is a hint to help bindings in providing optional arguments in
generated apis (e.g. languages with function overloading).
It should be used only for simple types like string, integer or float.
(Other types would need fancy parsing that is beyond the purpose of
this file)
Note: Arguments providing this tag are still required on the wire,
which leaves the protocol unchanged from the server point of view.