OTHER: Let --with-custom-version set full version.
Lets not restrict the imagination of packagers and let the
parameter replace the full version string. The parameter will
be used as a format string and {"version": .., "commit": ...}
will be passed to it with the current base version, and the
commit hash with -dirty appended if the workspace has changed.