OTHER: Use PYTHONARCHDIR for xmmsclient module.
Based on multiple conclusions over the web, for example:

"A python package must reside entirely in one directory. PEP 382
proposed support for namespace-packages that could be split in
different directories, but it was rejected. (And in any case,
those would be special packages.)

Given that python packages have to be in a single directory,
it is not possible to mix compiled extension modules for
different architectures."