Python client's PropDict class doesn't handle map objects correctly
The PropDict class (which inherits from dict) assumes that the given "srcs" are a list.  However, these srcs are created using map(), which in python2 returned a list, but in python3 returns a map object.  This wouldn't be too bad, except unfortunately, the func in the map isn't idempotent, which means that iterating self._sources multiple times leads to bizarre results, ie. the key is found the first time, but not subsequent times.  This breaks clients that happen to look up multiple keys (or keys multiple times).

Since the list of sources is always small (AFAIK), the simplest solution is to just materialise it back into being a list at init time.