xmmsapi.pyx: fix compatibility with python3 style C callbacks
Without the change `xmms2` python bindings fails to build as:

    Error compiling Cython file:
    ------------------------------------------------------------
    ...
                            raise TypeError("Type '%s' is not callable" % cb.__class__.__name__)
                    self._cb = cb
                    if cb is not None and not self._cb_issetup:
                            self.result_tracker = rt
                            rt.track_result(self)
                            xmmsc_result_notifier_set_full(self.res, ResultNotifier, <void *> self, ResultDestroyNotifier)
                                                ^
    ------------------------------------------------------------
    src/clients/lib/python/xmmsapi.pyx:180:44: Cannot assign type 'bool (xmmsv_t *, void *) except? -1' to 'xmmsc_result_notifier_t'

Here `cython` compains that C function callbacks defined in python
code have a chance throwing python exceptions (which will make
them hard to inwind when crossing C boundary).

Normally C callbacks are not allowed to throw the exceptions if the C
code is not explicitly aware of python interpreter.

Python code is already careful not to propagate python exceptions.
Let's mark all the callbacks as `noexcept`.