clients/nycli/utils.c: Fix buffer overflow
format_url() is only assigning 255 bytes for the rpath, but the path will
be expanded by realpath() which can return a sring up to a maximum of
PATH_MAX bytes. And, so as a result, if long path names are used or
while creating playlists with multiple files we get a coredump with the
error:

*** buffer overflow detected ***: terminated
Aborted (core dumped)

Lets use PATH_MAX for rpath length so that we have buffer for the
maximum return from realpath().

Signed-off-by: Sudip Mukherjee <sudipm.mukherjee@gmail.com>